part 4

+++ core/api/users.py
@@ -12,24 +12,41 @@ api_users = Blueprint('api_users', __name__)
 @auth.auth_required()
 def users():
     limit = get_limit(api_users)
+    countus = UserRepository().getCountUsers()
+    total = countus['counter']
+    order_by = request.args.get('orderby',type=int)
 
     if limit > 100:
-        limit = 100
+        limit = 1000
+
+    print(total)
 
     rows = UserRepository().getAll([limit])
+    if order_by == 1:
+        print("ASC")
+        return {'response': 'ASC'}
+    elif order_by == 2:
+        print("DESC")
+        return {'response': 'DESC'}
+    elif order_by == 3:
+        print("Order by Username")
+        return {'response': 'Order by Username'}
+    elif order_by == 4:
+        print("Order by tgid")
+        return {'response': 'Order by tgid'}
+    else:
+        return jsonify(list(map(lambda row: {
+            'id': row['id'],
+            'tg_id': row['tg_id'],
+            'username': row['tg_username'],
+            'warn': row['warn_count'],
+        }, rows)))
 
-    return jsonify(list(map(lambda row: {
-        'id': row['id'],
-        'tg_id': row['tg_id'],
-        'username': row['tg_username'],
-        'warn': row['warn_count'],
-    }, rows)))
-
-@api_users.route('/users/<int:tg_id>', methods=['GET'])
+@api_users.route('/users/search/tgid/<int:tg_id>', methods=['GET'])
 @limiter.limit("5000 per day")
 @limiter.limit("10/seconds")
 @auth.auth_required()
-def user(tg_id):
+def user_by_id(tg_id):
     row = UserRepository().getById([tg_id])
     if row:
         return {'id': row['id'],
@@ -39,11 +56,29 @@ def user(tg_id):
     else:
         return {'error': 'You have entered an id that does not exist or you have entered incorrect data'}
 
+@api_users.route('/users/search/username/<username>', methods=['GET'])
+@limiter.limit("5000 per day")
+@limiter.limit("10/seconds")
+@auth.auth_required()
+def user_by_username(username):
+    row = UserRepository().getByUsername(username)
+    if row:
+        return {'id': row['id'],
+        'tg_id': row['tg_id'],
+        'username': row['tg_username'],
+        'warn': row['warn_count']}
+    else:
+        return {'error': 'You have entered an username that does not exist or you have entered incorrect data'}
+
 @api_users.route('/users/delete_user/<int:tg_id>', methods=['DELETE'])
 @limiter.limit("500 per day")
 @limiter.limit("2/seconds")
 @auth.auth_required()
 def delete_user(tg_id):
-    data = [(tg_id)]
-    UserRepository().deleteUser(data)
-    return {'response': 'User successfully deleted'}
+    row = UserRepository().getById([tg_id])
+    if row:
+        data = [(tg_id)]
+        UserRepository().deleteUser(data)
+        return {'response': 'User successfully deleted'}
+    else:
+        return {'response': 'There is no user with this id to delete'}
